[WIP] Refactor some native ETH libs
