+++ consensus/indexed.go
@@ -3,7 +3,8 @@ package consensus
 import (
 	"math/big"
 
-	"github.com/unicornultrafoundation/go-u2u/libs/common"
+	"github.com/unicornultrafoundation/go-u2u/common"
+
 	"github.com/unicornultrafoundation/go-hashgraph/consensus/dagidx"
 	"github.com/unicornultrafoundation/go-hashgraph/hash"
 	"github.com/unicornultrafoundation/go-hashgraph/native/dag"
