+++ hash/event_hash.go
@@ -9,7 +9,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/unicornultrafoundation/go-u2u/libs/common"
+	"github.com/unicornultrafoundation/go-u2u/common"
+
 	"github.com/unicornultrafoundation/go-hashgraph/common/bigendian"
 	"github.com/unicornultrafoundation/go-hashgraph/native/idx"
 )
