+++ native/pos/validators.go
@@ -6,7 +6,7 @@ import (
 	"math"
 	"sort"
 
-	"github.com/unicornultrafoundation/go-u2u/libs/rlp"
+	"github.com/unicornultrafoundation/go-u2u/rlp"
 
 	"github.com/unicornultrafoundation/go-hashgraph/native/idx"
 )
