+++ types/cheaters_list.go
@@ -1,7 +1,8 @@
 package types
 
 import (
-	"github.com/unicornultrafoundation/go-u2u/libs/rlp"
+	"github.com/unicornultrafoundation/go-u2u/rlp"
+
 	"github.com/unicornultrafoundation/go-hashgraph/native/idx"
 )
