+++ u2udb/multidb/records.go
@@ -1,7 +1,8 @@
 package multidb
 
 import (
-	"github.com/unicornultrafoundation/go-u2u/libs/rlp"
+	"github.com/unicornultrafoundation/go-u2u/rlp"
+
 	"github.com/unicornultrafoundation/go-hashgraph/u2udb"
 )
