+++ vecengine/vecflushable/backed_map.go
@@ -1,7 +1,8 @@
 package vecflushable
 
 import (
-	"github.com/unicornultrafoundation/go-u2u/libs/common"
+	"github.com/unicornultrafoundation/go-u2u/common"
+
 	"github.com/unicornultrafoundation/go-hashgraph/u2udb"
 )
