+++ vecengine/vecflushable/vecflushable.go
@@ -3,7 +3,8 @@ package vecflushable
 import (
 	"errors"
 
-	"github.com/unicornultrafoundation/go-u2u/libs/common"
+	"github.com/unicornultrafoundation/go-u2u/common"
+
 	"github.com/unicornultrafoundation/go-hashgraph/u2udb"
 )
